pyproject.toml
[project]
name = "ai-tester"
version = "0.1.0"
description = "AI-based adaptive test scenario generation for backend systems (master's thesis prototype)"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Ayxan Şahsuvarov" },
]
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.5",
    "requests>=2.31",
    "gymnasium>=0.29",
    "stable-baselines3>=2.2",
    "numpy>=1.26",
    "google-generativeai>=0.5",
]

[project.optional-dependencies]
experiments = [
    "matplotlib>=3.8",
]

[project.scripts]
ai-tester = "ai_tester.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ai_tester", "sample_backend"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint.per-file-ignores]
"experiments/*.py" = ["E402"]

[dependency-groups]
dev = [
    "ruff>=0.15.12",
]
